fix(autoinstrumentation): DD_SERVICE in some init containers inconsistent
Addresses: https://datadoghq.atlassian.net/browse/INPLAT-550

This isn't a complete fix since we would also want to remove
the code that sets the environment variable based on the "owner
name" at some point (the source of this confusion).

This makes it so that the likelyhood of inconsistent env-vars
when this is called is minimized and it's more obvious what
the behavior does (by not tying it to lib-config).

Changes:

1. Moves the implicit service naming out of `libConfig` so this
   behavior is less hidden and more obvious that it's happening.

2. Adds behavior to find an already set `corev1.EnvVar` for `DD_SERVICE`
   and use that if it is found before setting one based on the owner
   name.

Behavior changes:

- For users where `DD_SERVICE` was set by `tagsFromLabels` this would
  make the init containers injected consistent with the application
  container. `tagsFromLabels` can use the labels on the pod OR the
  labels on the owner, and if thats set they would propagate as strings.

- For users where it was not set, this would find the owner name and
  then propagate it.