system-probe: Increase check API timeout
The low timeouts (5s) in the general API make sense if the client is
used by some code which needs to query system-probe for something as
part of some other task, but when the client is used by an agent check
which needs the data from system-probe as its main source of
information, the timeouts could be too low, so increase them to avoid
spurious errors.

The low timeouts are still used for the startup check since that
endpoint is not expected to take large amounts of time.